requirements.txt
farm-haystack[inference]
flask
trafilatura
requests
python-dotenv
transformers
pandas